requirements.txt
django==3.2.7
django-annoying==0.10.6
django-extensions
django-localflavor
djangorestframework==3.12.4
ipython
psycopg2==2.9.1
psycopg2-binary==2.9.1